[HM-588] Only register webhook when necessary
This change modifies NestGram to ask Telegram if a webhook is currently
registered for the bot. If the bot is already registered, the webhook
does not try to re-register itself.

This prevents a thundering-herd problem caused by many instances bot
starting up at once, all attempting to register the webhook, though not
completely -- in this situation, it's likely that many bots will still
fail to register an endpoint due to rate-limiting, but once they start
back up they won't need to re-register.